refactor(frontend): serve healthz only from root routes
