+++ frontend/providers/applaunchpad/next.config.js
@@ -39,11 +39,6 @@ const nextConfig = {
   async rewrites () {
     return [
       {
-        source: '/healthz',
-        destination: '/api/healthz',
-        locale: false
-      },
-      {
         source: '/api/v2alpha/docs',
         destination: '/doc/v2alpha'
       },
