+++ frontend/providers/costcenter/src/utils/healthz.ts
@@ -1,0 +1,7 @@
+import { getAppConfigResponse, initAppConfig } from '@/pages/api/platform/getAppConfig';
+export const HEALTHZ_SERVICE = 'costcenter';
+
+export function assertReady() {
+  initAppConfig();
+  getAppConfigResponse();
+}
