+++ frontend/providers/license/next.config.js
@@ -22,15 +22,6 @@ const nextConfig = {
     // this includes files from the monorepo base two directories up
     outputFileTracingRoot: path.join(__dirname, '../../')
   },
-  async rewrites() {
-    return [
-      {
-        source: '/healthz',
-        destination: '/api/healthz',
-        locale: false
-      }
-    ];
-  }
 };
 
 module.exports = nextConfig;
