+++ frontend/providers/objectstorage/next.config.js
@@ -71,15 +71,6 @@ module.exports = (phase, { defaultConfig }) => {
           }
         ];
       },
-      async rewrites() {
-        return [
-          {
-            source: '/healthz',
-            destination: '/api/healthz',
-            locale: false
-          }
-        ];
-      },
       transpilePackages: ['@sealos/ui', 'sealos-desktop-sdk']
     };
     return nextConfig;
@@ -91,14 +82,5 @@ module.exports = (phase, { defaultConfig }) => {
       },
       output: 'standalone',
       transpilePackages: ['@sealos/ui', 'sealos-desktop-sdk'],
-      async rewrites() {
-        return [
-          {
-            source: '/healthz',
-            destination: '/api/healthz',
-            locale: false
-          }
-        ];
-      }
     };
 };
