+++ frontend/providers/template/next.config.js
@@ -36,15 +36,6 @@ const nextConfig = {
     ]
   },
   transpilePackages: ['@sealos/ui', 'sealos-desktop-sdk', '@sealos/driver'],
-  async rewrites() {
-    return [
-      {
-        source: '/healthz',
-        destination: '/api/healthz',
-        locale: false
-      }
-    ];
-  }
 };
 
 module.exports = withBundleAnalyzer(nextConfig);
