+++ frontend/providers/template/src/utils/healthz.ts
@@ -1,0 +1,6 @@
+import { getClientAppConfigServer } from '@/pages/api/platform/getClientAppConfig';
+export const HEALTHZ_SERVICE = 'template';
+
+export function assertReady() {
+  getClientAppConfigServer();
+}
